re #17 Adding functions to replace ReadAsArray calls
Adding two functions in utilities.py for handling raster needs. One is a summation function that calls pygeoprocessing iterblocks and the other returns the min / max values.

There is still a call in mesh.py to ReadAsArray, for loading the map into matplotlib to show. Need to consult with Justin to see best way to handle this.