re #5 adding file dialogue for user to select save path
Although currently MESH has a very strict file structure, I have added a dialogue window that allows the user to specify where they would like to save their PDF report.

The path default that is brought up in the window is what was hardcoded in for the HTML save action.