re #2 updating function to check for sub directory existence.
refactored the function that validates if base data is good to go by
checking for certain sub directories and that they are not empty.

It is pretty easy to add sub directories to the list, and does not require
a CSV to map out all the data files, which would turn into a maintenance task.

File IO errors related to base data should be handled with a dialogue window to the user