re #28 Creating directory where json file should be placed.
The JSON file is designated to be saved in the workspace directory for the model run.

Since that directory is not created until the invest model is run (which invest creates it), we create it first now, and save the json file to that location.

NOTE: if this becomes a problem we can save and look for the json files in a higher level directory, I don't think it needs to be saved in this location