re #29 Updating how an InVEST run is validated
InVEST run is now validated if the *latest* model log file has succeeded
AND a json archive parameter file is present.

The json archive parameter file is generated from the InVEST UI and saves the current args_id based on UI selections. This is key to be able to properly launch an InVEST model from a script. Without this file we have no way of knowing how to properly set up the "args_id" dictionary to be passed in to execute function.