re #30 patch for modifying invest args
Was missing the case where we just want to use the defaults as set in the setup json files.
This catches that case, and also adds in some currently sketchy logic to see if
the default is a file path, in which case we want to join it with the base data folder.
If not a file path, then just set as is (float, bool, etc...)